# backend/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Amie API",
    "version": "1.0.0",
    "description": "Amie 애플리케이션의 API 문서"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "개발 서버"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/chat-rooms/user/{userId}": {
      "get": {
        "summary": "특정 사용자의 모든 채팅방 조회",
        "tags": [
          "ChatRooms"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "사용자 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "채팅방 목록",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "chatRooms": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string"
                          },
                          "user1Id": {
                            "type": "string"
                          },
                          "user2Id": {
                            "type": "string"
                          },
                          "isActive": {
                            "type": "boolean"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          },
          "404": {
            "description": "사용자를 찾을 수 없음"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
...
    "/api/credits/user/{userId}": {
      "get": {
        "summary": "특정 사용자의 크레딧 로그 조회",
        "tags": [
          "CreditLogs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "사용자 ID"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "페이지 번호"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "페이지당 항목 수"
          }
        ],
        "responses": {
          "200": {
            "description": "크레딧 로그 목록",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "totalLogs": {
                      "type": "integer",
                      "example": 45
                    },
                    "totalPages": {
                      "type": "integer",
                      "example": 3
                    },
                    "currentPage": {
                      "type": "integer",
                      "example": 1
                    },
                    "logs": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string"
                          },
                          "userId": {
                            "type": "string"
                          },
                          "action": {
                            "type": "string",
                            "enum": [
                              "CHARGE",
                              "MATCH",
                              "PROFILE_UNLOCK"
                            ]
                          },
                          "amount": {
                            "type": "number"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    },
                    "currentCredit": {
                      "type": "number",
                      "example": 120
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          },
          "404": {
            "description": "사용자를 찾을 수 없음"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
...
    "/api/credit/use": {
      "post": {
        "summary": "크레딧 사용",
        "tags": [
          "Credits"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount",
                  "service"
                ],
                "properties": {
                  "amount": {
                    "type": "number",
                    "description": "사용할 크레딧 양",
                    "example": 10
                  },
                  "service": {
                    "type": "string",
                    "description": "서비스 유형",
                    "example": "match"
                  },
                  "description": {
                    "type": "string",
                    "description": "사용 내역 설명",
                    "example": "매칭 서비스 이용"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "크레딧 사용 성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "인증 실패"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
...
    "/api/users": {
      "post": {
        "summary": "새 사용자 생성",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password",
                  "nickname"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "securePassword123"
                  },
                  "nickname": {
                    "type": "string",
                    "example": "사용자닉네임"
                  },
                  "birthYear": {
                    "type": "number",
                    "example": 1990
                  },
                  "height": {
                    "type": "number",
                    "example": 175
                  },
                  "city": {
                    "type": "string",
                    "example": "Seoul"
                  },
                  "gender": {
                    "type": "string",
                    "enum": [
                      "male",
                      "female"
                    ],
                    "example": "male"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "사용자 생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "user": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "인증 실패"
          },
          "409": {
            "description": "중복된 이메일 또는 닉네임"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      },
      "get": {
        "summary": "모든 사용자 목록 조회",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "페이지 번호"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "페이지당 사용자 수"
          }
        ],
        "responses": {
          "200": {
            "description": "사용자 목록",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "totalUsers": {
                      "type": "integer",
                      "example": 50
                    },
                    "totalPages": {
                      "type": "integer",
                      "example": 3
                    },
                    "currentPage": {
                      "type": "integer",
                      "example": 1
                    },
                    "users": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string"
                          },
                          "nickname": {
                            "type": "string"
                          },
                          "gender": {
                            "type": "string"
                          },
                          "birthYear": {
                            "type": "number"
                          },
                          "height": {
                            "type": "number"
                          },
                          "city": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "summary": "특정 사용자 정보 조회",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "사용자 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "사용자 정보",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "nickname": {
                          "type": "string"
                        },
                        "gender": {
                          "type": "string"
                        },
                        "birthYear": {
                          "type": "number"
                        },
                        "height": {
                          "type": "number"
                        },
                        "city": {
                          "type": "string"
                        },
                        "profileImages": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "credit": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          },
          "404": {
            "description": "사용자를 찾을 수 없음"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      },
      "put": {
        "summary": "사용자 정보 수정",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "사용자 ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nickname": {
                    "type": "string",
                    "example": "변경된닉네임"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "newPassword123"
                  },
                  "birthYear": {
                    "type": "number",
                    "example": 1992
                  },
                  "height": {
                    "type": "number",
                    "example": 178
                  },
                  "city": {
                    "type": "string",
                    "example": "Busan"
                  },
                  "profileImages": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "image1.jpg",
                      "image2.jpg"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "사용자 정보 수정 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "사용자 정보가 업데이트되었습니다."
                    },
                    "user": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "인증 실패"
          },
          "404": {
            "description": "사용자를 찾을 수 없음"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      },
      "delete": {
        "summary": "사용자 삭제",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "사용자 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "사용자 삭제 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "사용자가 삭제되었습니다."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          },
          "404": {
            "description": "사용자를 찾을 수 없음"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
    "/api/users/{id}/credit": {
      "post": {
        "summary": "사용자 크레딧 추가",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "사용자 ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount"
                ],
                "properties": {
                  "amount": {
                    "type": "number",
                    "description": "추가할 크레딧 양",
                    "example": 50
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "크레딧 추가 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "크레딧이 추가되었습니다."
                    },
                    "credit": {
                      "type": "number",
                      "example": 150
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "인증 실패"
          },
          "404": {
            "description": "사용자를 찾을 수 없음"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    }
  },
...
}